+++ odoo/songs/updates/update_10_1_1/clean_translation.sql
@@ -7,3 +7,10 @@ WHERE
     module = 'sale_discount_program'
 AND lang = 'fr_FR'
 AND src = 'Quantity is the sum of matching products quantities';
+
+DELETE FROM
+    ir_translation
+WHERE
+    module = 'specific_base'
+AND lang = 'fr_FR'
+AND src = 'Company name';