+++ odoo/local-src/specific_base/wizard/__init__.py
@@ -2,5 +2,4 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import models
-from . import wizard
+from . import crm_lead_to_opportunity